# Phase-2/specs/003-frontend-integration/contracts/frontend-api-contracts.yaml
openapi: 3.0.0
info:
  title: Frontend Integration API Contracts
  version: 1.0.0
  description: API contracts for frontend-backend integration in the Todo application

paths:
  /api/auth/login:
    post:
      summary: User login with Better Auth
      description: Authenticate user credentials and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful, JWT token returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Bad request - Invalid input data
        '401':
          description: Unauthorized - Invalid credentials
        '500':
          description: Internal server error

  /api/auth/register:
    post:
      summary: User registration with Better Auth
      description: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Bad request - Invalid input data or email already exists
        '500':
          description: Internal server error

  /api/auth/profile:
    get:
      summary: Get authenticated user profile
      description: Retrieve profile information for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized - Invalid or expired token
        '500':
          description: Internal server error

  /api/tasks:
    get:
      summary: Retrieve all tasks for the authenticated user
      description: Returns a list of all tasks associated with the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful response with user's tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing authentication
        '500':
          description: Internal server error
    post:
      summary: Create a new task for the authenticated user
      description: Creates a new task associated with the authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input data
        '401':
          description: Unauthorized - Invalid or missing authentication
        '500':
          description: Internal server error

  /api/tasks/{task_id}:
    get:
      summary: Retrieve a specific task
      description: Returns a specific task if it belongs to the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          description: ID of the task to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Successful response with the task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing authentication
        '404':
          description: Task not found or does not belong to the user
        '500':
          description: Internal server error
    put:
      summary: Update a specific task
      description: Updates a specific task if it belongs to the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          description: ID of the task to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input data
        '401':
          description: Unauthorized - Invalid or missing authentication
        '404':
          description: Task not found or does not belong to the user
        '500':
          description: Internal server error
    delete:
      summary: Delete a specific task
      description: Deletes a specific task if it belongs to the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          description: ID of the task to delete
          schema:
            type: string
      responses:
        '200':
          description: Task deleted successfully
        '401':
          description: Unauthorized - Invalid or missing authentication
        '404':
          description: Task not found or does not belong to the user
        '500':
          description: Internal server error

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          format: password
          description: User's password

    LoginResponse:
      type: object
      required:
        - access_token
        - token_type
        - user
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          description: Token type (bearer)
        user:
          $ref: '#/components/schemas/UserProfile'

    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          format: password
          description: User's password
        name:
          type: string
          description: User's name (optional)

    RegisterResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'

    UserProfile:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: string
          description: Unique identifier for the user
        email:
          type: string
          format: email
          description: User's email address
        name:
          type: string
          description: User's name (optional)
        created_at:
          type: string
          format: date-time
          description: When the user account was created

    Task:
      type: object
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique identifier for the task
        title:
          type: string
          description: Title of the task
          maxLength: 255
        description:
          type: string
          description: Detailed description of the task
        completed:
          type: boolean
          description: Whether the task is completed
          default: false
        user_id:
          type: string
          description: ID of the user who owns this task
        created_at:
          type: string
          format: date-time
          description: Timestamp when task was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when task was last updated

    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Title of the task
          maxLength: 255
        description:
          type: string
          description: Detailed description of the task

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          description: Updated title of the task
          maxLength: 255
        description:
          type: string
          description: Updated description of the task
        completed:
          type: boolean
          description: Updated completion status of the task